requirements.txt
awscli==1.10.0
boto==2.39.0
botocore==1.3.22
colorama==0.3.3
cycler==0.9.0
dj-database-url==0.3.0
<<<<<<< HEAD
Django==1.9
=======
Django==1.9.1
>>>>>>> origin/master
django-storages==1.1.8
docutils==0.12
gunicorn==19.4.5
jmespath==0.9.0
matplotlib==1.5.0
numpy==1.10.1
Pillow==3.0.0
psycopg2==2.6.1
pyasn1==0.1.9
pyparsing==2.0.6
python-dateutil==2.4.2
pytz==2015.7
rsa==3.3
six==1.10.0
virtualenv==13.1.2
whitenoise==2.0.6
